backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
cryptography>=41.0.0
django-ratelimit>=4.1.0
openai>=1.0.0
ollama>=0.1.0
celery>=5.3.0
django-celery_beat>=2.5.0
redis>=5.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
python-telegram-bot>=20.7
pandas>=2.0.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
ruff
drf-nested-routers
django-allauth
beautifulsoup4
xlsxwriter>=3.0.5
openpyxl>=3.1.0
python-docx>=0.8.11